Revamped docker build pipeline
I revamped the build pipeline in this commit to work in a more complex way.
Some of the complexity is for efficiency's sake (such as building different architectures in parallel).
Since all the images need to be present in the registry itself before you can use docker manifest create,
I decided to put the creation of the manifest and the tagging of images in its own job that depends on build-push.